refactor: restructure backend codebase for improved organization and maintainability
- Removed StopDetail schema and stops endpoint from the API documentation.
- Refactored app.js to use a dedicated stops router.
- Moved database connection logic to a new db.js file.
- Created index.js for centralized configuration loading.
- Implemented StopController for handling stop-related requests.
- Added loaders for initializing database and static data.
- Created seeder for populating the database with bus data.
- Introduced BusRepository for bus-related database operations.
- Created stops.routes.js for defining stop-related routes.
- Removed seedBus.js as its functionality is now integrated into the seeder.
- Refactored StopService to handle fetching and processing stop details.
- Updated swagger.js to reflect new configuration structure.
- Added staticData.js for loading stops and routes from JSON files.
