--- backend/docs/components/schemas/StopDetail.yaml
@@ -1,77 +1,0 @@
-components:
-  schemas:
-    StopDetail:
-      type: object
-      properties:
-        distance:
-          type: number
-          nullable: true
-          description: "Always Null"
-        routes:
-          type: array
-          description: "List of routes that pass through this stop"
-          items:
-            type: object
-            properties:
-              areaId:
-                type: integer
-                description: "It is the Id of the area the route belongs to. Urban: 23 for Trento, 24 for Rovereto, 22 for l'Alto Garda, 21 for Pergine Valsugana; Extra-urban: 1, 2, 3, 4, 5, 6; Railway: 7; Cableway: 8"
-              news:
-                type: string
-                description: "If there is any news related to this route, like changes of the route due to works or events"
-                nullable: true
-              routeColor:
-                type: string
-                example: "#00A88F"
-                description: "Color associated with the route"
-                nullable: true
-              routeId:
-                type: integer
-                description: "Id of the route"
-              routeLongName:
-                description: "Full name of the route"
-                type: string
-              routeShortName:
-                description: "Short name of the route"
-                type: string
-              routeType:
-                type: integer
-                description: "Type of the route: U = Urban, E = Extra-Urban"
-              type:
-                type: string
-        stopCode:
-          description: "Code of the stop"
-          type: string
-        stopDesc:
-          description: "Description of the stop"
-          type: string
-        stopId:
-          description: "Id of the stop"
-          type: integer
-        stopLat:
-          description: "Latitude of the stop"
-          type: number
-          format: float
-        stopLevel:
-          description: "Altitude of the stop"
-          type: integer
-        stopLon:
-          description: "Longitude of the stop"
-          type: number
-          format: float
-        stopName:
-          description: "Name of the stop"
-          type: string
-        street:
-          description: "Street where the stop is located"
-          type: string
-        town:
-          description: "Town where the stop is located"
-          type: string
-          nullable: true
-        type:
-          description: "Type of the stop: U=Urban, E=Extra-Urban"
-          type: string
-        wheelchairBoarding:
-          description: "Indicates whether the stop is wheelchair accessible: 1 = Accessible, 0 and 2 = Not accessible"
-          type: integer
