+++ backend/src/config/db.js
@@ -1,0 +1,21 @@
+import { connect } from "mongoose";
+import config from "./index.js";
+
+export async function connectDb() {
+  const { user, pass, host, name } = config.db;
+
+  if (!user || !pass || !host) {
+    console.error("❌ Error: Missing MongoDB environment variables");
+    process.exit(1);
+  }
+
+  const mongoURI = `mongodb://${user}:${pass}@${host}:27017/${name}?authSource=admin`;
+
+  try {
+    await connect(mongoURI);
+    console.log("✅ Connected to MongoDB successfully!");
+  } catch (err) {
+    console.error("❌ Error connecting to MongoDB:", err.message);
+    process.exit(1);
+  }
+}
