+++ backend/src/routes/stops.routes.js
@@ -1,0 +1,8 @@
+import express from "express";
+import { getStop } from "../controllers/StopController.js";
+
+const router = express.Router();
+
+router.get("/:id", getStop);
+
+export default router;
