+++ backend/src/server.js
@@ -1,19 +1,22 @@
-import { app, connectDb } from "./app.js";
-import { initData } from "./data.js";
-import config from "./config.js";
-import { seedDatabase } from "./seedBus.js";
+import { app } from "./app.js";
+import config from "./config/index.js";
+import loaders from "./loaders/index.js";
 
 const port = config.api.port;
 
 async function startServer() {
   try {
-    await Promise.all([connectDb(), initData(), seedDatabase()]);
+    await loaders();
 
     app.listen(port, () => {
-      console.log(`Server API listening on port ${port}`);
+      console.log(`
+      ################################################
+      🛡️  Server listening on port: ${port} 🛡️
+      ################################################
+      `);
     });
   } catch (err) {
-    console.error("Failed to start server:", err);
+    console.error("❌ Failed to start server:", err);
     process.exit(1);
   }
 }
